backend/requirements.txt
Django==5.0.6
djangorestframework==3.15.2
django-cors-headers==4.4.0
djangorestframework-simplejwt==5.3.1
python-dotenv==1.0.1
# psycopg2-binary==2.9.9  # Commenté - utilisé uniquement si USE_POSTGRESQL=1
